added --no-watcher arg

+++ src/index.ts
@@ -8,14 +8,14 @@ import {onProcessExit, ThrottleTime} from "./utils";
 
 function printUsage() {
     console.log('Usage: `cd` into your root package folder first, then execute');
-    console.log('npm-local-development <package1-name> <package1-source> <package2-name> <package2-source> ...');
+    console.log('npm-local-development <package1-name> <package1-source> <package2-name> <package2-source> ... [--no-watcher]');
     console.log('or to use Lerna');
-    console.log('npm-local-development lerna');
+    console.log('npm-local-development lerna [--no-watcher]');
     console.log('or to read config .sync.json:');
-    console.log('npm-local-development config');
+    console.log('npm-local-development config [--no-watcher]');
 }
 
-async function sync(cwd: string, packageName: string, packageSource: string) {
+async function sync(cwd: string, packageName: string, packageSource: string, watching = true) {
     const ignored: string[] = [];
 
     const packageJsonPath = join(cwd, 'package.json');
@@ -60,16 +60,19 @@ async function sync(cwd: string, packageName: string, packageSource: string) {
 
     readDeps();
 
-    onProcessExit(async () => {
-        closed = true;
-        for (const watcher of watchers) {
-            watcher.close();
-        }
+    if (watching) {
+        //we reset the state back when we had a watcher running
+        onProcessExit(async () => {
+            closed = true;
+            for (const watcher of watchers) {
+                watcher.close();
+            }
 
-        log(`Exiting, reverting directory structure for ${rootPackageName} (${packagePathInRootNodeModules})`);
-        await fs.remove(packagePathInRootNodeModules);
-        await fs.ensureSymlink(packageSource, packagePathInRootNodeModules);
-    });
+            log(`Exiting, reverting directory structure for ${rootPackageName} (${packagePathInRootNodeModules})`);
+            await fs.remove(packagePathInRootNodeModules);
+            await fs.ensureSymlink(packageSource, packagePathInRootNodeModules);
+        });
+    }
 
     if (await fs.pathExists(packagePathInRootNodeModules)) {
         await fs.remove(packagePathInRootNodeModules);
@@ -127,43 +130,46 @@ async function sync(cwd: string, packageName: string, packageSource: string) {
     }
 
     await updateNodeModulesSymLinks();
-    const throttledUpdateNodeModulesSymLinks = ThrottleTime(() => updateNodeModulesSymLinks(), 1);
 
-    /**
-     * What for changes in the origin package source, .e.g. '../core/package.json', this is important
-     * to re-read peerDependencies.
-     */
-    watchers.push(watch(packageSource + '/package.json', {
-        ignoreInitial: true,
-        ignored: ['.git'],
-        followSymlinks: false
-    }).on('all', (event, path) => {
-        log('package.json changed, reload.');
-        readDeps();
-        throttledUpdateNodeModulesSymLinks();
-    }));
+    if (watching) {
+        const throttledUpdateNodeModulesSymLinks = ThrottleTime(() => updateNodeModulesSymLinks(), 1);
+
+        /**
+         * What for changes in the origin package source, e.g. '../core/package.json', this is important
+         * to re-read peerDependencies.
+         */
+        watchers.push(watch(packageSource + '/package.json', {
+            ignoreInitial: true,
+            ignored: ['.git'],
+            followSymlinks: false
+        }).on('all', (event, path) => {
+            log('package.json changed, reload.');
+            readDeps();
+            throttledUpdateNodeModulesSymLinks();
+        }));
 
 
-    /**
-     * Watch for changes in origin package source, e.g. '../core/', so we can copy
-     * files manually to our root package's node_modules/{packageName}/
-     */
-    watchers.push(watch(packageSource, {
-        ignoreInitial: true,
-        followSymlinks: false
-    }).on('all', async (event, path) => {
-        if (path.startsWith(resolve(join(packageSource, 'node_modules')))) return;
+        /**
+         * Watch for changes in origin package source, e.g. '../core/', so we can copy
+         * files manually to our root package's node_modules/{packageName}/
+         */
+        watchers.push(watch(packageSource, {
+            ignoreInitial: true,
+            followSymlinks: false
+        }).on('all', async (event, path) => {
+            if (path.startsWith(resolve(join(packageSource, 'node_modules')))) return;
 
-        const target = join(packagePathInRootNodeModules, relative(packageSource, path));
+            const target = join(packagePathInRootNodeModules, relative(packageSource, path));
 
-        // log(event, relative(cwd, path), '->', relative(packageSource, path));
+            // log(event, relative(cwd, path), '->', relative(packageSource, path));
 
-        if (event === 'unlink') {
-            fs.unlink(target);
-        } else {
-            fs.copy(path, target);
-        }
-    }));
+            if (event === 'unlink') {
+                fs.unlink(target);
+            } else {
+                fs.copy(path, target);
+            }
+        }));
+    }
 }
 
 async function run() {
@@ -171,6 +177,9 @@ async function run() {
         printUsage();
         return;
     }
+
+    const watching = process.argv.filter(v => v === '--no-watcher').length === 0;
+    process.argv = process.argv.filter(v => v !== '--no-watcher');
 
     const cwd = process.cwd();
     if ('lerna' === process.argv[2]) {
@@ -218,7 +227,7 @@ async function run() {
 
                 for (const depToSync in depsToSync) {
                     console.log(`${chalk.green(packages[pkg])} -> ${chalk.green(depToSync)} (${depsToSync[depToSync]})`);
-                    promises.push(sync(join(cwd, packages[pkg]), depToSync, join(cwd, depsToSync[depToSync])));
+                    promises.push(sync(join(cwd, packages[pkg]), depToSync, join(cwd, depsToSync[depToSync]), watching));
                 }
             } catch (error) {
                 throw new Error(`Could not read package.json of ${path}`);
@@ -250,7 +259,7 @@ async function run() {
                 const packageName = i;
                 const packageSource = syncConfig[cwd][i];
                 console.log(`  ${chalk.green(packageName)} -> ${chalk.green(packageSource)}`);
-                promises.push(sync(cwd, packageName, packageSource));
+                promises.push(sync(cwd, packageName, packageSource, watching));
             }
         }
 
@@ -269,7 +278,7 @@ async function run() {
 
         for (const i in packages) {
             console.log(`${chalk.green(i)} -> ${chalk.green(packages[i])}`);
-            promises.push(sync(cwd, i, packages[i]));
+            promises.push(sync(cwd, i, packages[i], watching));
         }
 
         await Promise.all(promises);
